+++ zoom/_assets/standard_apps/sample/fields.py
@@ -40,7 +40,7 @@ form1 = Form(
     PasswordField('Password', hint='test hint'),
     NumberField('Number of Things', hint='test hint'),
     NumberField('Area', units='ft<sup>2</sup>', hint='test hint'),
-    # IntegerField('Count', hint='test hint'),
+    IntegerField('Count', hint='test hint'),
     FloatField('Size', units='meters', hint='test hint'),
     DecimalField('Decimal', hint='test hint'),
     MoneyField('Amount', hint='test hint'),