+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add created_by and updated_by to records created by User.add method
 
 ### Added
+- fix bug in IntegerField zero values
 
 ## [6.15.0] - 2021-02-06
 - add installed path to admin environment info page